Fix tests controllers, conditions for User, Article and Category
