+++ templates/article/show.html.twig
@@ -46,7 +46,7 @@
                             <h3><i class="pr-3 fa fa-comment"></i>{{ article.comments|length }} {{ 'comments'|trans }}</h3>
                             <hr>
 
-                            {% if is_granted('ROLE_USER') %}
+                            {% if is_granted('IS_AUTHENTICATED_FULLY') %}
                                     <div class="row mb-5">
                                         <div class="col-sm-12">
                                             {{ render(controller('App\\Controller\\Web\\ArticleController::commentForm', {'id': article.id})) }}
@@ -78,15 +78,17 @@
 {% endblock %}
 
 {% block sidebar %}
-    <div class="section">
-        <a href="{{ path('article_edit', {slug: article.slug}) }}" class="btn btn-lg btn-block btn-success">
-            <i class="fa fa-edit" aria-hidden="true"></i> {{ 'action.edit_contents'|trans }}
-        </a>
-    </div>
+    {% if article.isAuthor(app.user) %}
+        <div class="section">
+            <a href="{{ path('article_edit', {slug: article.slug}) }}" class="btn btn-lg btn-block btn-success">
+                <i class="fa fa-edit" aria-hidden="true"></i> {{ 'action.edit_contents'|trans }}
+            </a>
+        </div>
 
-    <div class="section">
-        {{ include('article/_delete_form.html.twig', {article: article}, with_context = false) }}
-    </div>
+        <div class="section">
+            {{ include('article/_delete_form.html.twig', {article: article}, with_context = false) }}
+        </div>
+    {% endif %}
 
     {{ render(controller('App\\Controller\\Web\\HomeController::sidebar')) }}
