+++ templates/profile/index.html.twig
@@ -14,7 +14,7 @@
                     <div class="col-md-6 no-pad">
                         <div class="user-pad">
                             <h3>{{ 'profile.wellcome_back'|trans }} {{ user.fullName }}</h3>
-                            <a class="btn btn-labeled btn-info" href="{{ path('user_edit', {id: user.id}) }}">
+                            <a class="btn btn-labeled btn-info" href="{{ path('user_edit') }}">
                                 <span class="btn-label"><i class="fa fa-pencil"></i></span>{{ 'button.update'|trans }}
                             </a>
                         </div>
