+++ templates/user/edit.html.twig
@@ -10,8 +10,6 @@
     {{ form_row(form.fullName) }}
     {{ form_row(form.username) }}
     {{ form_row(form.email) }}
-    {{ form_row(form.password.first) }}
-    {{ form_row(form.password.second) }}
     {{ form_end(form) }}
 
 {% endblock %}
