+++ tests/Controller/Web/DefaultControllerTest.php
@@ -6,11 +6,11 @@ use App\Entity\Article;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\HttpFoundation\Response;
 
-
 class DefaultControllerTest extends WebTestCase
 {
     /**
      * @dataProvider getPublicUrls
+     * @param string $url
      */
     public function testPublicUrls(string $url)
     {
@@ -28,13 +28,14 @@ class DefaultControllerTest extends WebTestCase
         $client = static::createClient();
         // the service container is always available via the test client
         $article = $client->getContainer()->get('doctrine')->getRepository(Article::class)->find(1);
-        $client->request('GET', sprintf('/en/article/edit/%s', $article->getSlug()));
+        $client->request('GET', sprintf('/en/article/%s', $article->getSlug()));
         $this->assertSame(Response::HTTP_OK, $client->getResponse()->getStatusCode());
     }
 
     /**
      * @dataProvider getSecureUrls
-    */
+     * @param string $url
+     */
     public function testSecureUrls(string $url)
     {
         $client = static::createClient();
@@ -42,7 +43,7 @@ class DefaultControllerTest extends WebTestCase
         $response = $client->getResponse();
         $this->assertSame(Response::HTTP_FOUND, $response->getStatusCode());
         $this->assertSame(
-            'http://127.0.0.1:8000/en/login',
+            'http://localhost/en/login',
             $response->getTargetUrl(),
             sprintf('The %s secure URL redirects to the login form.', $url)
         );
@@ -57,9 +58,8 @@ class DefaultControllerTest extends WebTestCase
 
     public function getSecureUrls()
     {
-        yield ['/en/article/list-articles'];
+        yield ['/en/article/user-articles'];
         yield ['/en/article/new'];
-        yield ['/en/article/1'];
-        yield ['/en/article/1/edit'];
+        yield ['/en/article/edit/art1'];
     }
 }
